refactor: extract shared types into a dedicated module
